Move build into own folder
