+++ vitest.config.mjs
@@ -1,7 +1,7 @@
 import swc from 'unplugin-swc';
 import { defineConfig } from 'vite';
 
-import { viteTestConfig, swcTestConfig } from './build.mjs';
+import { viteTestConfig, swcTestConfig } from './build/test.mjs';
 
 export default defineConfig({
   test: viteTestConfig,
